qdrant-qa-bot/requirements.txt
qdrant-client
sentence-transformers
gpt4all
requests
markdown
beautifulsoup4
python-telegram-bot